AB Test to Randomize cookies and links
This code is used to conduct an AB test of two
different application forms for DIV. Users will
be provided one of two possible forms at random.
Users who have visited the page before will always
get the same form based on a randomly selected cookie
set in their browser. (Given the informal nature of the
experiment, this code does not cover all possible ways
that a user could access the form.) When DIV receives the
application forms, we will conduct an AB test to compare
the answers provided in the applications.